functions/requirements.txt
# Production dependencies
firebase_functions~=0.1.0
firebase-admin>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Development dependencies (install locally with: pip install -r requirements-dev.txt)
# These are separated to avoid bloating production functions